+++ frameworks/react/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to the library will be documented in this file.
 
+## vX.X.X (Month DD, YYYY)
+
+- Change imports from `@formisch/core/vanilla` to `@formisch/core/react`
+- Change imports from `@formisch/methods/vanilla` to `@formisch/methods/react`
+
 ## v0.2.2 (December 12, 2025)
 
 - Change `@formisch/core` to v0.4.5